[x64] Factor out vector handling code
